backend/server/requirements.txt
werkzeug
Flask
Flask-SQLAlchemy
psycopg2-binary
pandas
numpy
yfinance
apscheduler
secure-smtplib
openai
binance-futures-connector==4.0.0
binance-connector==3.5.1
tzlocal==4.2
PySide6==6.3.0
pyqtgraph==0.12.4
qdarkstyle==3.1
matplotlib==3.5.3
seaborn==0.11.2
ta-lib==0.4.24
deap==1.3.3
pyzmq==23.2.1
plotly==5.10.0
importlib-metadata==4.12.0
tqdm==4.64.1
peewee==3.17.0
backports.zoneinfo; python_version < '3.9'